refactor(170): reorganize domain puzzle model tests
