+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/model/StripEntryRangeTest.kt
@@ -1,6 +1,5 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs.domain.puzzle.model
 
-import org.cescfe.numpairs.domain.puzzle.model.StripEntryRange
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertThrows
 import org.junit.Assert.assertTrue
