+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/model/StripTest.kt
@@ -1,9 +1,5 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs.domain.puzzle.model
 
-import org.cescfe.numpairs.domain.puzzle.model.Strip
-import org.cescfe.numpairs.domain.puzzle.model.StripEntry
-import org.cescfe.numpairs.domain.puzzle.model.StripEntryRange
-import org.cescfe.numpairs.domain.puzzle.model.StripItem
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertThrows
 import org.junit.Test
